refactor: move error/context to appropriate sub-modules
